8281879: Serial: Merge CardGeneration into TenuredGeneration
