--- src/hotspot/share/gc/shared/cardGeneration.inline.hpp
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2014, 2019, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- *
- */
-
-#ifndef SHARE_GC_SHARED_CARDGENERATION_INLINE_HPP
-#define SHARE_GC_SHARED_CARDGENERATION_INLINE_HPP
-
-#include "gc/shared/cardGeneration.hpp"
-
-#include "gc/shared/space.hpp"
-
-inline size_t CardGeneration::capacity() const {
-  return space()->capacity();
-}
-
-inline size_t CardGeneration::used() const {
-  return space()->used();
-}
-
-inline size_t CardGeneration::free() const {
-  return space()->free();
-}
-
-inline MemRegion CardGeneration::used_region() const {
-  return space()->used_region();
-}
-
-inline bool CardGeneration::is_in(const void* p) const {
-  return space()->is_in(p);
-}
-
-inline CompactibleSpace* CardGeneration::first_compaction_space() const {
-  return space();
-}
-
-#endif // SHARE_GC_SHARED_CARDGENERATION_INLINE_HPP
